fix: consolidate files to satisfy pkg-file-count CI limit
Merge workflow validation apisurface helpers into factory_preview.go,
fold validation tests into factory_preview_test.go, and move terminology
CLI tests into root_session_test.go so pkg/apisurface and pkg/cli stay
within the 15-file package limit.
